% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condaDownloadStats.R
\name{anacondaDownloadStats}
\alias{anacondaDownloadStats}
\title{Get download statistics for Bioconductor packages distributed via
Anaconda.}
\usage{
anacondaDownloadStats()
}
\value{
A \code{data.frame} of download statistics for
all Bioconductor packages distributed by Anaconda, in tidy format.
Note: Anaconda do not provide counts for unique IP addresses.  This column
is listed as \code{NA} for all packages to provide continuity with data from
Bioconductor.org obtained by \code{\link[=biocDownloadStats]{biocDownloadStats()}}.  The counts are
updated monthly, so do not expect to see counts for the current month.
}
\description{
Get download statistics for Bioconductor packages distributed via
Anaconda.
}
\details{
Anaconda provide daily download counts for all software packages
they distribute.  These are summarised into monthly tables of counts and made
available from https://github.com/grimbough/anaconda-download-stats
This function provides a mechanism to download these monthly counts for
Bioconductor packages distributed through Anaconda.
}
\examples{
anacondaDownloadStats()

}
\author{
Mike L. Smith
}
