% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiveInstall.R
\name{archiveInstall}
\alias{archiveInstall}
\title{Install packages from a previous release of Bioconductor for reproducibility}
\usage{
archiveInstall(
  pkgs = character(),
  version = BiocManager::version(),
  snapshot = getOption("BiocPkgTools.snapshot", "RSPM"),
  ...
)
}
\arguments{
\item{pkgs}{\code{character()} vector of package names to install or
update.  A missing value updates installed packages according
to \verb{update =} and \verb{ask =}. Package names containing a '/' are
treated as GitHub repositories and installed using
\code{remotes::install_github()}.}

\item{version}{character(1) The desired version to reproduce. This is largely
dictated by the current R / Bioconductor version installed and is indicated
by \code{BiocManager::version} by default.}

\item{snapshot}{character(1) The snapshot CRAN repository to use for
reproducibility. This defaults to the value of
\code{getOption("BiocPkgTools.snapshot", "RSPM")}.}

\item{...}{Additional arguments used by \code{install.packages()}.}
}
\value{
Mostly called for the side-effects of copying and modifying the
\code{config.yaml} and \code{.Renviron} files to reproduce an R / Bioconductor
package environment from a previous Bioconductor release.
}
\description{
This function allows users to install packages from a previously released
Bioconductor version.
}
\details{
The \code{version} argument is largely dictated by the actual R /
Bioconductor installation, e.g., Bioconductor version 3.14. With this
information and the value of \code{getOption("BiocPkgTools.snapshot")}, the
function will temporarily change the \code{getOption('repos')} setting for
\code{CRAN} to allow installation of CRAN packages from either the RSPM or MRAN
time machines. The date of the 'snapshot' coincides with the \emph{last} release
date for the indicated Bioconductor version. The function will also modify
the \code{BIOCONDUCTOR_USE_CONTAINER_REPOSITORY} environment variable to
temporarily disable binary package installations.
}
\examples{

if (interactive()) {
  archiveInstall("DESeq2", version = "3.14")
}

}
