% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocConfigVersion.R
\name{biocConfigVersion}
\alias{biocConfigVersion}
\title{Create a Bioconductor configuration that allows reproducibility}
\usage{
biocConfigVersion(
  version = "3.15",
  lastBuildDate = NULL,
  snapshot = getOption("BiocManager.snapshot", "RSPM"),
  yamlDir = tools::R_user_dir("BiocPkgTools", which = "config"),
  overwrite = FALSE,
  Renviron = file.path(Sys.getenv("HOME"), ".Renviron")
)
}
\arguments{
\item{version}{character(1) The desired version to reproduce. This is largely
dictated by the current R / Bioconductor version installed.}

\item{lastBuildDate}{named character(1) Override argument to specify the date
of the desired CRAN snapshot, this date should agree or come close to the
date provided by \code{biocLastBuildDate}. The name of the character vector
should correspond to the Bioconductor version, e.g.,
\code{c('3.15'="2022-10-19")}. When \code{lastBuildDate} is not given, the \code{version}
argument is used in conjunction with the aforementioned function.}

\item{snapshot}{character(1) The snapshot CRAN repository to use for
reproducibility. This defaults to the value of
\code{getOption("BiocManager.snapshot", "RSPM")}.}

\item{yamlDir}{character(1) The folder location for the modified
\code{config.yaml} file. By default, it is set to the value of
\code{tools::R_user_dir("BiocPkgTools", which = "config")}.}

\item{overwrite}{logical(1) Whether to overwrite the local \code{config.yaml}
file. Set to \code{FALSE} by default. Set to \code{TRUE} in order to save the
changes made to the configuration map.}

\item{Renviron}{character(1) The file location of the \code{.Renviron} file to
update with the \code{BIOCONDUCTOR_CONFIG_FILE} variable. Set to \code{NULL} to avoid
writing in the \code{.Renviron}. The default directory location of the file is
given by \code{Sys.getenv("HOME")}.}
}
\value{
Mostly called for the side-effects of copying and modifying the
\code{config.yaml} and \code{.Renviron} files to reproduce an R / Bioconductor
package environment from a previous Bioconductor release.
}
\description{
This function helps users who wish to use a previously released Bioconductor
version.
}
\details{
The \code{version} argument is largely dictated by the current R /
Bioconductor installation, e.g., Bioconductor version 3.15. With this
information and the value of \code{getOption("BiocManager.snapshot")}, the
function will download a copy of the \code{config.yaml} file and modify it with
the appropriate date entry as returned from \code{biocLastBuildDate} for the
Bioconductor version in question. The function will also update the
\code{.Renviron} file with a \code{BIOCONDUCTOR_CONFIG_FILE} entry. See
\code{?BiocManager::install} for more details. This step can be skipped by
setting the \code{Renviron} argument to \code{NULL} if the \code{BIOCONDUCTOR_CONFIG_FILE}
option is set.
}
\examples{
biocConfigVersion("3.15", "RSPM")
}
