% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocExplore.R
\name{biocExplore}
\alias{biocExplore}
\title{Explore Bioconductor packages interactively}
\usage{
biocExplore(top = 500L, ...)
}
\arguments{
\item{top}{maximum number of packages displayed in any biocView}

\item{...}{parameters passed to \code{htmlwidgets::createWidget()}}
}
\value{
A bubble plot of Bioconductor packages
}
\description{
Explore Bioconductor packages through an interactive bubble plot. Click on bubbles to bring up additional information about the package. Size and proximity to center of a bubble is based on the downloads the package has in the past month.
}
