% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocLastBuildDate.R
\name{biocLastBuildDate}
\alias{biocLastBuildDate}
\title{Obtain the last build date for a particular Bioconductor version}
\usage{
biocLastBuildDate(version = "all")
}
\arguments{
\item{version}{character(1) Indicates the Bioconductor version for which the
last build date is sought. By default, 'all' versions will be returned.}
}
\description{
The function facilitates the discovery of last build dates useful for
selecting a fixed date to be used in conjunction with
\code{options(BiocPkgTools.snapshot)}. Currently, it looks at
\url{https://bioconductor.org/checkResults/} and parses the dates listed.
}
\examples{

biocLastBuildDate(version = "3.14")

}
