% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocLastBuildDate.R
\name{BiocPkgTools-defunct}
\alias{BiocPkgTools-defunct}
\alias{biocLastBuildDate}
\title{Deprecated functions in \code{BiocPkgTools}}
\usage{
biocLastBuildDate(version)
}
\arguments{
\item{version}{character(1) Indicates the Bioconductor version for which the
last build date is sought.}
}
\description{
Currently, \code{biocLastBuildDate} is \strong{DEPRECATED}. See
functionality in \code{BiocArchive}.
}
\details{
The function facilitates the discovery of last build dates useful
for selecting a fixed date. Currently, it looks at
\url{https://bioconductor.org/checkResults/} and parses the dates listed.
}
\seealso{
\url{https://github.com/LiNk-NY/BiocArchive}
}
