% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_status.R
\name{build_report}
\alias{build_report}
\title{Tidy Bioconductor build report results}
\usage{
build_report(version = biocVersion())
}
\arguments{
\item{version}{character(1) the version number
as used to access the online build report. For 
example, "3.6".}
}
\value{
a data.frame with columns pkg, version,
author, commit, date, node, stage, and result.
}
\description{
The online Bioconoductor build reports
are great for humans to look at, but
they are not easily computable. This function
simply scrapes HTML and text files available
from the build report online pages to generate
a tidy data frame version of the build report.
}
\examples{

latest_build = build_report("3.6")
head(latest_build)

}
