% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDependencies.R
\name{class-dependencies}
\alias{class-dependencies}
\alias{classDepTree}
\alias{classDepData}
\title{Retrieve Class relationships}
\usage{
classDepTree(class, includeUnions = FALSE)

classDepData(class, includeUnions = FALSE)

classDepData(class, includeUnions = FALSE)
}
\arguments{
\item{class}{a single \code{character} value defining a \sQuote{S4} class
name}

\item{includeUnions}{\code{TRUE} or \code{FALSE}: Should union definitions
included in the result? (default: \code{FALSE})}
}
\description{
As the title says it should do something with class relationships
}
\examples{
depData <- classDepData("RangedSummarizedExperiment")
classDepData("RangedSummarizedExperiment")
}
