% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocDownloadStats.R
\name{commitsSince}
\alias{commitsSince}
\title{How many commits have created since a date?}
\usage{
commitsSince(local_repo, Date, exclude_author)
}
\arguments{
\item{local_repo}{character(1) The path to a local git repository for which
to review the commit history}

\item{Date}{character(1) The date cutoff from which to analyze closed issues
in the year-month-day format.}

\item{exclude_author}{character() A vector of author names to exclude from
the results}
}
\value{
A \code{tibble} with commit history from the \verb{git log} command
}
\description{
The \code{commitsSince} function allows maintainers to count the number of commits
since a given date using the \verb{git log} command.
}
\details{
The \code{tibble} given by the function has 5 columns, 'datetime',
'commit', 'parents', 'author', and 'subject'. The 'parents' column refers
to the parent commit.
}
\examples{
if (interactive()) {
    commitsSince("~/bioc/S4Vectors", "2021-05-05")
}

}
