% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataciteXMLGenerate.R
\name{dataciteXMLGenerate}
\alias{dataciteXMLGenerate}
\title{The Bioconductor datacite.org XML generator}
\usage{
dataciteXMLGenerate(pkg)
}
\arguments{
\item{pkg}{The name of a Bioconductor package}
}
\value{
An xml_document object from the \code{xml2} package.
}
\description{
This function is used internally to generate XML elements from the
datacite.org website for incoming Bioconductor packages.
}
\seealso{
\preformatted{?xml2::`xml_document-class`}
}
\keyword{internal}
