% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocPkgList.R
\name{.extract_fnd}
\alias{.extract_fnd}
\title{Extract funder names from an \code{Authors@R} field value}
\usage{
.extract_fnd(authors_at_r_string)
}
\arguments{
\item{authors_at_r_string}{character(1) the value of the \code{Authors@R}
DESCRIPTION field}
}
\value{
character() of funder names, or \code{NA_character_} when no funder
is found or parsing fails
}
\description{
Extract funder names from an \code{Authors@R} field value
}
\keyword{internal}
