% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_orcid.R
\name{.get_cre_orcid}
\alias{.get_cre_orcid}
\title{get the ORCID id from cre field of Authors@R in packageDescription result}
\usage{
.get_cre_orcid(pkgname)
}
\arguments{
\item{pkgname}{character(1)}
}
\description{
get the ORCID id from cre field of Authors@R in packageDescription result
}
