% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_orcid.R
\name{.get_orcid_rec}
\alias{.get_orcid_rec}
\title{process employment data from ORCID}
\usage{
.get_orcid_rec(orcid, rename = TRUE)
}
\arguments{
\item{orcid}{character(1)}

\item{rename}{logical(1) if TRUE use short names}
}
\description{
process employment data from ORCID
}
