% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_bioc_data}
\alias{get_bioc_data}
\title{Get data from bioconductor}
\usage{
get_bioc_data()
}
\value{
json string containing bioconductor package details
}
\description{
Get data from bioconductor
}
\examples{
bioc_data <- get_bioc_data()
}
