% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocDownloadStats.R
\name{getBiocDownloadStats}
\alias{getBiocDownloadStats}
\title{get bioconductor download stats}
\usage{
getBiocDownloadStats()
}
\value{
a \code{data.frame} of download stats for
all bioconductor packages, in tidy format
}
\description{
get bioconductor download stats
}
\details{
Note that bioconductor package download
stats are not version-specific.
}
