% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiocPkgList.R
\name{getBiocPkgList}
\alias{getBiocPkgList}
\title{Get full bioc software package listing}
\usage{
getBiocPkgList(version = biocVersion())
}
\arguments{
\item{version}{The requested bioconductor version. Will
default to use the BiocInstaller defaults (ie., \code{biocVersion()}).}
}
\value{
a \code{data.frame}
}
\description{
The BiocViews-generated \code{VIEWS} file is available
for bioconductor release and devel repositories. It
contains quite a bit more information from the
package \code{DESCRIPTION} files than the \code{PACKAGES}
file. In particular, it contains \code{biocViews} annotations
and URLs for vignettes and developer URLs.
}
