% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiocViewsFile.R
\name{getBiocViewsFile}
\alias{getBiocViewsFile}
\title{Get full biocViews file}
\usage{
getBiocViewsFile(version = biocVersion())
}
\arguments{
\item{biocVersion}{The requested bioconductor version. Will
default to use the BiocInstaller defaults (ie., \code{biocVersion()}).}
}
\value{
a \code{data.frame}
}
\description{
The BiocViews-generated \code{VIEWS} file is available
for bioconductor release and devel repositories. It
contains quite a bit more information from the
package \code{DESCRIPTION} files than the \code{PACKAGES}
file. In particular, it contains \code{biocViews} annotations
and URLs for vignettes and developer URLs.
}
