% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackageInfo.R
\name{getPackageInfo}
\alias{getPackageInfo}
\title{Generate needed information to create DOI from a package
directory.}
\usage{
getPackageInfo(dir)
}
\arguments{
\item{dir}{\code{character(1)} Path to package}
}
\value{
A \code{data.frame}
}
\description{
Generate needed information to create DOI from a package
directory.
}
\keyword{Internal}
