% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignetteTools.R
\name{getVignette}
\alias{getVignette}
\title{Get vignette}
\usage{
getVignette(vignettePath, destfile = tempfile(),
  biocVersion = BiocManager::version())
}
\arguments{
\item{vignettePath}{the additional path information to get to the vignette}

\item{destfile}{the file location to store the vignette}

\item{biocVersion}{defaults to user version}
}
\description{
Get vignette
}
\examples{
\dontrun{
x = getBiocPkgList()
tmp = getVignette(x$vignettes[[1]][1])
library(pdftools)
y = pdf_text(tmp)
y = paste(y,collapse=" ")
library(tm)
v = VCorpus(VectorSource(y))
library(magrittr)

v <- v \%>\%
    tm_map(stripWhitespace) \%>\%
    tm_map(content_transformer(tolower)) \%>\%
    tm_map(removeWords, stopwords("english")) \%>\%
    tm_map(stemDocument)
dtm = DocumentTermMatrix(v)
inspect(DocumentTermMatrix(v,
    list(dictionary = as.character(x$Package))))
}
}
