% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubUtils.R
\name{githubURLParts}
\alias{githubURLParts}
\title{Extract github user and repo name from github URL}
\usage{
githubURLParts(urls)
}
\arguments{
\item{urls}{a \code{character()} vector of urls.}
}
\value{
a \code{data.frame} with four columns:
  \itemize{
    \item{url}{The original github URL}
    \item{user_repo}{The github "username/repo", combined}
    \item{user}{The github username}
    \item{repo}{The github repo name}
  }
}
\description{
Extract github user and repo name from github URL
}
\examples{
# find github URL details for
# Bioc packages
bpkgl = biocPkgList()
urldetails = githubURLParts(bpkgl$URL)
urldetails = urldetails[!is.na(urldetails$url),]
head(urldetails)

}
