% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocDownloadStats.R
\name{issuesSince}
\alias{issuesSince}
\title{What are the issues created since a date?}
\usage{
issuesSince(
  gh_repo,
  Date,
  status = c("closed", "open", "all"),
  issue_metadata = c("created_at", "number", "title")
)
}
\arguments{
\item{gh_repo}{character(1) The GitHub repository location including the
username / organization and the repository name, e.g.,
"Bioconductor/S4Vectors"}

\item{Date}{character(1) The date cutoff from which to analyze closed issues
in the year-month-day format.}

\item{status}{character(1) One of 'closed', 'open', or 'all' corresponding to
the issue state desired from the GitHub API (Default: "closed")}

\item{issue_metadata}{character() The metadata labels to extract from the
\code{gh::gh} response. See \code{?gh::gh} for more details. Defaults to
'created_at', 'number', and 'title'.}
}
\value{
A \code{tibble} with three columns corresponding to issue metadata (i.e.,
"created_at", "number", "title")
}
\description{
This function uses the \code{gh} package to get a list of issues since the
specified date for a particular GitHub repository. The repository must
have both the username / organization and the name, e.g.,
"Bioconductor/S4Vectors".
}
\examples{

issuesSince("Bioconductor/S4Vectors", "2021-05-01", "closed")

}
