% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_orcid.R
\name{orcid_table}
\alias{orcid_table}
\title{Obtain employment data from ORCID}
\usage{
orcid_table(orcids)
}
\arguments{
\item{orcids}{\code{character()} A vector of ORCID identifiers}
}
\value{
a \code{data.frame} of employment info using the ORCID API
}
\description{
Get a \code{data.frame} of employment info from ORCID
}
\examples{
if (interactive()) {
    orcid_table(
        orcids = c(
            "0000-0002-3242-0582",
            "0000-0003-4046-0063",
            "0000-0003-2725-0694"
        )
    )
}
}
