% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocBuildEmail.R
\name{biocBuildEmail}
\alias{biocBuildEmail}
\title{Create and copy e-mail package notification template to clipboard}
\usage{
biocBuildEmail(pkg, version = c("release", "devel"),
  emailTemplate = .getTemplatePath(), core.name = NULL,
  core.email = NULL)
}
\arguments{
\item{pkg}{The name of the package in trouble}

\item{version}{character(1) the version number
as used to access the online build report. For 
example, "3.6". The default is the "current version"
as specified in \code{BiocManager::version}. Note
that this is a character() variable, not a number.}

\item{emailTemplate}{The path to the email template.}
}
\value{
a character string of the email
}
\description{
The \code{biocBuildEmail} function provides a template for notifying
maintainers of errors in the Bioconductor Build System (BBS). This
convenience function returns the body of the email from a template
within the package and provides a copy in the clipboard.
}
