% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocDownloadStats.R
\name{biocDownloadStats}
\alias{biocDownloadStats}
\title{Get Bioconductor download statistics}
\usage{
biocDownloadStats(
  pkgType = c("software", "data-experiment", "workflows", "data-annotation")
)
}
\arguments{
\item{pkgType}{\code{character()} All, some, or one of 'software',
'data-experiment', 'workflows', or 'data-annotation' (defaults to all
types)}
}
\value{
A \code{tibble} of download statistics for all Bioconductor packages
}
\description{
Get Bioconductor download statistics
}
\details{
Note that Bioconductor package download
stats are not version-specific.
}
\examples{
biocDownloadStats()

}
