% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocDownloadStats.R
\name{biocDownloadStats}
\alias{biocDownloadStats}
\title{Get Bioconductor download statistics}
\usage{
biocDownloadStats(
  pkgType = c("all", "software", "data-experiment", "workflows", "data-annotation")
)
}
\arguments{
\item{pkgType}{character(1) Either one of 'software', 'data-experiment',
'workflows', or 'data-annotation' (defaults to 'all' or 'software')}
}
\value{
A \code{data.frame} of download statistics for
all Bioconductor packages, in tidy format
}
\description{
Get Bioconductor download statistics
}
\details{
Note that Bioconductor package download
stats are not version-specific.
}
\examples{
biocDownloadStats()

}
