% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocPkgList.R
\name{biocPkgList}
\alias{biocPkgList}
\title{Get full bioc software package listing, with details}
\usage{
biocPkgList(version = BiocManager::version(), repo = "BioCsoft")
}
\arguments{
\item{version}{The requested bioconductor version. Will
default to use the BiocManager defaults (ie., \code{version()}).}

\item{repo}{The requested biooconductor repository. The default will be the
Bioconductor software repository: BioCsoft. Available repos include:
"BioCsoft", "BioCann", "BioCexp", "BioCworkflows", and "CRAN". Note
that not all repos are available for all versions, particularly older
versions (but who would use those, right?).}
}
\value{
an object of class \code{tbl_df}.
}
\description{
The BiocViews-generated \code{VIEWS} file is available
for bioconductor release and devel repositories. It
contains quite a bit more information from the
package \code{DESCRIPTION} files than the \code{PACKAGES}
file. In particular, it contains \code{biocViews} annotations
and URLs for vignettes and developer URLs.
}
\examples{
bpkgl = biocPkgList()
bpkgl
unlist(bpkgl[1,'Depends'])

# Get a list of all packages that
# import "GEOquery"
library(dplyr)
bpkgl \%>\%
  filter(Package=='GEOquery') \%>\%
  pull(c('importsMe'))

}
