% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{BiocPkgTools-cache}
\alias{BiocPkgTools-cache}
\alias{setCache}
\alias{pkgToolsCache}
\title{Manage cache for BiocPkgTools}
\usage{
setCache(directory = rappdirs::user_cache_dir("BiocPkgTools"),
  verbose = TRUE, ask = interactive())

pkgToolsCache(...)
}
\arguments{
\item{directory}{The file location where the cache is located. Once set
future downloads will go to this folder.}

\item{verbose}{Whether to print descriptive messages}

\item{ask}{logical (default TRUE when interactive session) Confirm the file
location of the cache directory}

\item{...}{For `cbioCache`, arguments passed to `setCache`}

\item{cancer_study_id}{A single string from `studiesTable` associated
with a study tarball}
}
\description{
Managing user data is important to allow use of email functions
such as `biocBuildEmail` and made easy with `BiocFileCache`.
}
\section{pkgToolsCache}{

Get the directory location of the cache. It will prompt the user to create
a cache if not already created. A specific directory can be used via
\code{setCache}.
}

\section{setCache}{

Specify the directory location of the data cache. By default, it will
got to the user's home/.cache and "appname" directory as specified by
\link{user_cache_dir}. (default appname: cBioPortalData)
}

\section{removeCache}{

Some files may become corrupt when downloading, this function allows
the user to delete the tarball associated with a `cancer_study_id` in the
cache.
}

