% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocRevDepEmail.R
\name{biocRevDepEmail}
\alias{biocRevDepEmail}
\title{Notify downstream maintainers of changes in upstream package}
\usage{
biocRevDepEmail(
  pkg,
  which = c("strong", "most", "all"),
  PS = character(1L),
  version = BiocManager::version(),
  dry.run = TRUE,
  cc = NULL,
  emailTemplate = templatePath("revdepnote"),
  core.name = NULL,
  core.email = NULL,
  core.id = NULL,
  textOnly = FALSE,
  verbose = FALSE,
  credFile = "~/.blastula_creds"
)
}
\arguments{
\item{pkg}{character(1) The name of the package for whose reverse
dependencies are to be checked and notified.}

\item{which}{a character vector listing the types of
    dependencies, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.
    Character string \code{"all"} is shorthand for that vector,
    character string \code{"most"} for the same vector without
    \code{"Enhances"},
    character string \code{"strong"} (default) for the first three
    elements of that vector.
  }

\item{PS}{character(1) Postscript, an additional note to the recipient of
the email (i.e., the package maintainer)}

\item{version}{character() A vector indicating which version of Bioconductor
the package is failing in (either 'release' or 'devel'; defaults to both)}

\item{dry.run}{logical(1) Display the email without sending to the recipient.
It only works for HTML email reports and ignored when `textOnly=TRUE`}

\item{cc}{character() Email addresses for sending the message as a carbon
copy}

\item{emailTemplate}{character(1) The path to the email template Rmd file as
obtained by `templatePath()`. A custom template can be provided as file
path.}

\item{core.name}{character(1) The full name of the core team member}

\item{core.email}{character(1) The Roswell Park email of the core team
member}

\item{core.id}{character(1) The internal identifier for the Roswell employee.
This ID usually matches `^[A-Z]{2}[0-9]{5}` for more recent identifiers.}

\item{textOnly}{logical(1) Whether to return the text of the email only.
This avoids the use of the 'blastula' package and adds the text to the
system clipboard if the `clipr` package is installed (default: FALSE)}

\item{verbose}{logical(1) Whether to output full email information from
'smtp_send' (when `dry.run` is `FALSE` and 'blastula' is installed)}

\item{credFile}{character(1) An optional file generated by the
`blastula::create_smtp_creds_file` function containing email
 authentication information (default: "~/.blastula_creds"). See
 `?biocBuildEmail` details.}
}
\description{
The `biocRevDepEmail` function collects all the emails of the reverse
dependencies and sends a notification that an upstream package has been
deprecated or removed. It uses a template found in `inst/resources` with
the `templatePath()` function.
}
\examples{

biocRevDepEmail(
    "FindMyFriends", version = "3.13", dry.run = TRUE, textOnly = TRUE
)

}
