% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDependencies.R
\name{class-dependencies}
\alias{class-dependencies}
\alias{buildClassDepGraph}
\alias{buildClassDepData}
\alias{plotClassDep}
\alias{plotClassDepData}
\alias{plotClassDepGraph}
\alias{buildClassDepFromPackage}
\title{Retrieve Class relationships}
\usage{
buildClassDepGraph(class, includeUnions = FALSE)

buildClassDepData(class, includeUnions = FALSE)

plotClassDep(class, includeUnions = FALSE)

plotClassDepData(data)

plotClassDepGraph(g)

buildClassDepFromPackage(pkg)
}
\arguments{
\item{class}{a single \code{character} value defining a \sQuote{S4} class
name}

\item{includeUnions}{\code{TRUE} or \code{FALSE}: Should union definitions
included in the result? (default: \code{FALSE})}
}
\description{
As the title says it should do something with class relationships
}
\examples{
depData <- buildClassDepData("RangedSummarizedExperiment")
depData
g <- buildClassDepGraph("RangedSummarizedExperiment")
plotClassDepGraph(g)
}
