% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRANstatus.R
\name{CRANstatus}
\alias{CRANstatus}
\title{Check the CRAN build report page and email a notification}
\usage{
CRANstatus(pkg, core.email = NULL,
  to.mail = "maintainer@bioconductor.org",
  emailTemplate = .getTemplatePath("cranreport"))
}
\arguments{
\item{pkg}{character(1) The name of the package in trouble}

\item{core.email}{character(1) The Roswell Park email of the core team
member}

\item{to.mail}{The email of the CRAN report recipient}

\item{emailTemplate}{character(1) The path to the email template. The default
path lies in the 'inst' package folder.}
}
\description{
The \code{CRANstatus} function allows users to check the status of a package
and send an email report of any failures.
}
