% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{dataciteXMLGenerate}
\alias{dataciteXMLGenerate}
\title{Bioc datacite XML generator}
\usage{
dataciteXMLGenerate(pkg)
}
\arguments{
\item{pkg}{The name of a Bioconductor package}
}
\value{
An XML element
}
\description{
Bioc datacite XML generator
}
