% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{dataciteXMLGenerate}
\alias{dataciteXMLGenerate}
\title{Bioc datacite xml generator}
\usage{
dataciteXMLGenerate(pkg)
}
\arguments{
\item{pkg}{name of bioc package}
}
\value{
an xml element
}
\description{
Bioc datacite xml generator
}
