% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_bioc_data}
\alias{get_bioc_data}
\title{Get data from Bioconductor}
\usage{
get_bioc_data()
}
\value{
A JSON string containing Bioconductor package details
}
\description{
Get data from Bioconductor
}
\examples{
bioc_data <- get_bioc_data()
}
