% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_orcid.R
\name{get_cre_orcids}
\alias{get_cre_orcids}
\title{get ORCID ids from cre fields of Authors@R in packageDescription results}
\usage{
get_cre_orcids(pkgnames)
}
\arguments{
\item{pkgnames}{character() must be installed}
}
\description{
get ORCID ids from cre fields of Authors@R in packageDescription results
}
\note{
returns NA if no ORCID provided in Authors@R for package description
}
\examples{
get_cre_orcids(c("BiocPkgTools", "utils"))
}
