% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackageInfo.R
\name{getPackageInfo}
\alias{getPackageInfo}
\title{Generate needed inforamation to create DOI from a package
directory}
\usage{
getPackageInfo(dir)
}
\arguments{
\item{dir}{character(1) Path to package}
}
\value{
a \code{data.frame}
}
\description{
Generate needed inforamation to create DOI from a package
directory
}
\keyword{Internal}
