% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackageInfo.R
\name{getPackageInfo}
\alias{getPackageInfo}
\title{Generate needed inforamation to create DOI from a package
DESCRIPTION file}
\usage{
getPackageInfo(file)
}
\arguments{
\item{file}{character(1) Path to DESCRIPTION file}
}
\value{
a \code{data.frame}
}
\description{
Generate needed inforamation to create DOI from a package
DESCRIPTION file
}
