% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocDownloadStats.R
\name{latestPkgStats}
\alias{latestPkgStats}
\title{Summary of the latest package statistics}
\usage{
latestPkgStats(
  gh_repo,
  Date,
  pkgType = c("software", "data-experiment", "workflows", "data-annotation")
)
}
\arguments{
\item{gh_repo}{character(1) The GitHub repository location including the
username / organization and the repository name, e.g.,
"Bioconductor/S4Vectors"}

\item{Date}{character(1) The date cutoff from which to analyze closed issues
in the YYYY-MM-DD or YYYY-MM-DDTHH:MM:SSZ format (ISO 8601).}

\item{pkgType}{character(1) One of 'software', 'data-experiment',
'workflows', or 'data-annotation' (defaults to 'software')}
}
\description{
The \code{latestPkgStats} function combines outputs from several functions to
generate a table of relevant statistics for a given package.
}
\examples{
if (interactive()) {

  latestPkgStats("Bioconductor/BiocGenerics", "2021-05-05")

}
}
