% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_orcid.R
\name{orcid_table}
\alias{orcid_table}
\title{get data.frame of employment info from orcid}
\usage{
orcid_table(orcids)
}
\arguments{
\item{orcids}{character()}
}
\description{
get data.frame of employment info from orcid
}
\examples{
if (interactive()) {  # need a token?
oids = c("0000-0003-4046-0063", "0000-0003-4046-0063")
print(orcid_table(oids))
oids = c(oids, NA)
print(orcid_table(oids))
print(orcid_table(oids[1]))
}
}
