% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDependencyMetrics.R
\name{pkgDepImports}
\alias{pkgDepImports}
\title{Report package imported functionality}
\usage{
pkgDepImports(pkg)
}
\arguments{
\item{pkg}{character() name of the packge for which we want
to obtain metrics on its dependency burden.}
}
\value{
A tidy data frame with different metrics on the
package dependency burden.
}
\description{
Function adapted from 'itdepends::dep_usage_pkg' at https://github.com/r-lib/itdepends
}
