% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDependencyMetrics.R
\name{pkgDepMetrics}
\alias{pkgDepMetrics}
\title{Report package dependency burden}
\usage{
pkgDepMetrics(pkg, depdf)
}
\arguments{
\item{pkg}{character() name of the packge for which we want
to obtain metrics on its dependency burden.}

\item{depdf}{a tidy data frame with package dependency information
obtained through the function \code{\link{buildPkgDependencyDataFrame}}}
}
\value{
a tidy data frame with different metrics on the
package dependency burden.
}
\description{
Elaborate a report on the dependency burden of a given package.
}
