% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problemPage.R
\name{problemPage}
\alias{problemPage}
\title{generate hyperlinked HTML for problematic build reports for Bioc packages}
\usage{
problemPage(authorPatt = "V.*Carey", ver = "3.8")
}
\arguments{
\item{authorPatt}{character(1) regexp used with grep() to filter author field of package DESCRIPTION for listing}

\item{ver}{character(1) version tag for Bioconductor}
}
\value{
DT::datatable call; if assigned to a variable, must evaluate to get the page to appear
}
\description{
generate hyperlinked HTML for problematic build reports for Bioc packages
}
\examples{
if (interactive()) problemPage()
}
